tree_test/requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
drf-yasg==1.20.0
python-decouple==3.4
djangorestframework-recursive==0.1.2